requirements.txt
aiogram
asyncpg
hangul-romanize
gtts
python-dotenv
